// apps/api/src/types/enums.ts
export enum Role {
  RESIDENT = 'RESIDENT',
  STAFF = 'STAFF',
  ADMIN = 'ADMIN',
}

export enum Category {
  MAINTENANCE = 'MAINTENANCE',
  FOOD = 'FOOD',
  CLEANLINESS = 'CLEANLINESS',
  SAFETY = 'SAFETY',
  OTHER = 'OTHER',
}

export enum Status {
  OPEN = 'OPEN',
  IN_PROGRESS = 'IN_PROGRESS',
  RESOLVED = 'RESOLVED',
  CLOSED = 'CLOSED',
}